Update App.js
By using the same prop names as the default `button` and just proxying the props we care about, the `Button` component can used as a drop-in replacement for anything that uses the native `button` component. As-is, this change makes the `Button` component an extraneous proxy, but it does provide a single place to affect every button on the site as complexity grows. Personally, I like to wrap all user inputs in custom component, because they tend to be the most repeated and most consistent in the design and having that once place to change them all is super handy.

In `App`, replacing the explicit calls to `Button` for each increment with a mapped array, you poise the component to be easily refactored to have this data passed as a prop and be controlled by another component. If you had `prop-types` installed, I would've written it like this to have that prop already in place and just using `defaultProps` to set what should happen if the consumer doesn't care to customize.

```
import React from 'react';
import PropTypes from 'prop-types';

export default function App({ increments }) {
  const [counter, setCounter] = useState(0);
  const incrementCounter = (incrementValue) => setCounter(counter + incrementValue);
  return (
    <div>
    {increments.map(increment =>
        <Button
          key={increment}
          onClick={() => incrementCounter(increment)}
        >
          {toSignedString(increment)}
        </Button>
      )}
      <Display message={counter} />
    </div>
  );
}

App.propTypes = {
  increments: PropTypes.arrayOf(PropTypes.number),
};

App.defaultProps = {
  increments: [ -10, -5, -1, 1, 5, 10 ],
};
```